refactor: move fetch method into provider impl
